requirements.txt
# MCP Server Framework (FastMCP)
# Install from GitHub since it's not on PyPI yet
fastmcp @ git+https://github.com/jlowin/fastmcp.git

# HTTP client for API calls
httpx>=0.27.0

# Environment variables (optional, for local development)
python-dotenv>=1.0.0